fix(ui): enable auto scroll
The auto scroll functionality is added to the chat component to automatically scroll to the bottom of chat when new messages are added.

BREAKING CHANGE: The parameter name of the `useAutoScroll` hook has been changed from `ref` to `contentRef` to avoid conflicts with the `ref` prop of the `ChatContainer` component.